Working on d2input.go for #487 (#514)
d2input/ is now clear of lint.
